+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased]
 
+## [0.5.0] - 2025-07-13
+
 ### Added
 
 - Upgrade `cargo_metadata` to version 0.21
@@ -70,7 +72,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2025-07-09
 
 <!-- next-url -->
-[Unreleased]: https://github.com/bluurryy/cargo-insert-docs/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/bluurryy/cargo-insert-docs/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v0.5.0
 [0.4.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v0.4.0
 [0.3.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v0.3.0
 [0.2.1]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v0.2.1